Large refactor and fixed a http header bug
