+++ src/CaptainHook.Common/Constants.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CaptainHook.Common
+{
+    public struct Constants
+    {
+        public struct Headers
+        {
+            public const string CorrelationId = "X-Correlation-ID";
+        }
+    }
+}
