+++ src/CaptainHook.EventHandlerActor/Handlers/HttpFailureLogger.cs
@@ -10,7 +10,7 @@ namespace CaptainHook.EventHandlerActor.Handlers
     /// <summary>
     /// Wrapper for logging http failures in CH which include data re the endpoint being hit and the event from which the request came
     /// </summary>
-    [Obsolete]
+    [Obsolete("Need to wrap the http handlers in a polly rather than wrap the retries in a http client")]
     public class HttpFailureLogger
     {
         private readonly IBigBrother _bigBrother;
