+++ src/CaptainHook.EventHandlerActor/Handlers/IRequestLogger.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+using CaptainHook.Common;
+using CaptainHook.Common.Configuration;
+
+namespace CaptainHook.EventHandlerActor.Handlers
+{
+    public interface IRequestLogger
+    {
+        Task LogAsync(
+            HttpClient httpClient,
+            HttpResponseMessage response,
+            MessageData messageData,
+            Uri uri,
+            HttpVerb httpVerb
+        );
+    }
+}
