+++ src/Tests/CaptainHook.Tests/Web/Authentication/ContentTypeMatcher.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net.Http;
 using RichardSzalay.MockHttp;
 
-namespace CaptainHook.Tests.Authentication
+namespace CaptainHook.Tests.Web.Authentication
 {
     /// <inheritdoc />
     /// <summary>
