+++ src/Tests/CaptainHook.Tests/Web/Authentication/MmAuthenticationHandlerTests.cs
@@ -12,7 +12,7 @@ using Newtonsoft.Json;
 using RichardSzalay.MockHttp;
 using Xunit;
 
-namespace CaptainHook.Tests.Authentication
+namespace CaptainHook.Tests.Web.Authentication
 {
     public class MmAuthenticationHandlerTests
     {
